rename: domain 폴더 경로 수정
