+++ src/main/java/com/example/be_likelion_kwu_homepage/project/domain/ProjectCategory.java
@@ -1,4 +1,4 @@
-package com.example.be_likelion_kwu_homepage.domain;
+package com.example.be_likelion_kwu_homepage.project.domain;
 
 // 프로젝트 분류 Enum 클래스로 별도 관리
 public enum ProjectCategory {
